Unify layout legend data and keep vector symbol fallback
